requirements.txt
pandas
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
pydantic